+++ CHANGELOG.md
@@ -5,6 +5,14 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## v1.10.0
+
+### Changed
+
+- Backup the `wp_scripts` and `wp_styles` globals before any tests are run and
+  restore it before each HTTP request to prevent side effects from tests that
+  modify these globals.
+
 ## v1.9.5
 
 ### Added